Add oninput event to handle IBus insertions
The later part of handleAlpha() is separated into its own function
checkCharCorrect(). This way we can check for correct characters coming
both from keydown and input events.

Currently this doesn't count the characters in the textbox into all the
WPM stat magic. This can be fixed in the future by setting apart the
relevant chunks from keydown into separate functions to be called from (or
merged with) checkCharCorrect().

Also for some reason this only works on Chrome for now.