Handle IBus insertions by detecting Dead keys
Split functions again: took out a chunk from checkCharCorrect() and made
applyChar(). This so we can check if the character is correct before
applying its consequences. Specifically, we check from the oninput event.

Pressing a dead key or entering unicode insertion mode changes into
deadMode. Here, instead of sending all keypresses to applyChar() like
normal, we first check if they're correct. This happens until exiting
deadMode (either from oninput or because the right keypress was sent).

In oninput we do the same: once deadMode is true, we try to applyChar()
only if the character is correct.

deadMode is force-exited when pressing backspace, space or return. This to
avoid getting stuck inside it.

For now, this solution (ironically) only works 100% on Firefox, and needs
more testing.