Improvements to weakspot.
Gets rid of `defaultScore` (which was hacky), instead makes the learning rate faster at the start so that it's a simple mean for the first `perCharCount` of each letter and an exponential average afterwards. Chars that haven't been seen at all don't contribute to a word's score.

The effect should be that it seems "fairer" with only a small amount of input.