Update German Language Files
I noticed some of the German word being capitalised, when they should
not – unless at the start of the sentence. I believe this happened
because the sentences analysed where not normalized to fix this issue.

I believe that in monkeytype without punctuation the words should be
capitalised (or not) as if they were in the middle of a sentence. Only
if punctuation is enabled, words should get capitalized after a period.

To fix this I took a look at the original source mentioned in the
comment section of the JSON files. I wrote a script to download the
frequency map, correct capitalization, remove non-German words,
abbreviations and words from other languages.

Also, instead of only using one dataset I combined multiple ones to
reduce error and eliminate as much single occurrence words as possible.
This commit uses:
- wikipedia_2021_1M
- mixed-typical_2011_1M
- news_2021_1M
- wikipedia_2021_300K
- mixed-typical_2011_300K
- news_2021_300K

I do not believe this list of corrections to be exhaustive and it might
need adjustments in the future. Please feel free to open a PR against
the script I used: https://github.com/Syphdias/monkeytype-generate-german