Update `dynamic` Keymap Legend Style to More Accurately Reflect Casing (#4113) Ferotiq
* use `.on`

* update symbols and letters seperately

This is based off how Shift and Caps Lock are fundamentally different. Shift changes every key while Caps Lock just changes letters. In addition, MacOS and Windows handle Caps Lock differently. In Windows, Shift can revert the letters change of Caps Lock, and on MacOS it cannot. So I have also added in an edge case for MacOS.

* reduce the diff

* explain why there's a platform check

* remove export

* update `isUsingOppositeShift` to accept keycodes

This allows the keymap to update only what can be capitalized when `oppositeShiftMode` is on

* simplify

* fix support for iso layouts

* fix `oppositeShiftMode`: `keymap` not working

* should default to true

* whoops

* error handling