Update hebrew.json
replaced a few of the quotes.

id: 31
"They opened a hand grenade in their hearts: click! Ashkenazic? Iraqi? Brother! Jews! Warriors!" - famous quote that sounds better in Hebrew.

id: 32
"We have learned from history that wherever there is a train there is life" 

id: 36
"Officials in the police and the prosecutor's office teamed up with the left-wing newspapers, I call them 'only not Bibi', in order to sew up illusory cases for me. The goal is to topple a strong right-wing prime minister and thus distance the national camp from the country's leadership" 

translation to the other quotes is in the description of the first commit message