impr(quotes): add Hebrew quotes (#5262)
* Update hebrew.json

* Update hebrew.json

* Update hebrew.json

replaced a few of the quotes.

id: 31
"They opened a hand grenade in their hearts: click! Ashkenazic? Iraqi? Brother! Jews! Warriors!" - famous quote that sounds better in Hebrew.

id: 32
"We have learned from history that wherever there is a train there is life" 

id: 36
"Officials in the police and the prosecutor's office teamed up with the left-wing newspapers, I call them 'only not Bibi', in order to sew up illusory cases for me. The goal is to topple a strong right-wing prime minister and thus distance the national camp from the country's leadership" 

translation to the other quotes is in the description of the first commit message

* Update hebrew.json

fixed an error

* Update hebrew.json

updated some of the quotes.
here is the translation:

id: 31
"They opened a hand grenade in their hearts: click! Ashkenazic? Iraqi? Jews! Siblings! Warriors!"
-Menachem Begin



id: 32
"Words can lead to actions, but what can lead to actions is reality. Reality, not words!"
-Miri Regev

id: 33
"Honesty is strength, values are strength, cunning and corruption bring us down, generosity and love are the strongest forces in the world."
-Yair Lapid

id: 34
"The Jewish people did not excel at seeing dangers in time, time after time they surprised us - the last time in the most terrible way, and this will not happen under my leadership."
-Benjamin Netanyahu


id: 35
"Violence is eroding the foundation of Israeli democracy. She must be condemned, denounced, isolated. This is not the way of the State of Israel."
-Yitzhak Rabin

id: 36
"Factors in the police and the prosecutor's office teamed up with the left-wing newspapers, I call them 'just not Bibi', in order to sew up illusory cases for me. The goal is to topple a strong right-wing prime minister and thus distance the national camp from the country's leadership."
-Benjamin Netanyahu