feat: add quote from 'The Power Of Now'
Added a new quote to the database:

"Aren't past and future just as real, sometimes even more real, than the present? After all, the past determines who we are, as well as how we perceive and behave in the present. And our goals determine which actions we take in the present."

Source: 'The Power Of Now' (ID: 7615, Length: 239)