impr(quotes): add urdu quotes (@aitchz) (#6131)
### Description

In this pull request, I have added a set of Urdu quotes to contribute to
the collection of quotes in Monkeytype. These quotes are intended to
enhance the diversity of quotes available for typing practice in Urdu.

The quotes are properly formatted in JSON to fit the existing structure
used in the Monkeytype repository. Each quote includes its content and
the source, which is verified for authenticity.

### Translation

- Id 1: If you want people to take you seriously, stop laughing, and
then you'll take everything seriously.

- Id 2: A wise man sees his own faults and the world sees the faults of
a fool.

- Id 3: Always learn to compromise because giving in a little is better
than breaking a relationship forever.

- Id 4: Always make it easy for others in life. This world runs on the
principle that whatever a person does for others, he will get back more
than that. If we keep our hearts clean, others Help and correct your
actions, then believe that life will become a cradle of peace.

- Id 5: You must have heard, man is known by his intentions. But
remember, intentions are incomplete until they are transformed into
action. The secret of success is not only in planning but in executing
the plan. It's in getting dressed.

- Id 6: That which was in you in us, the song of conversation, that
mention, those dreams, those longings, those heartbeats that together
made a life, now they are dispersed in such a way that they cannot be
found. Time The current did not leave us anywhere, and took away from us
that moment, the scents that we had promised each other, the paths on
which we came closer to each other. Now we both realize that Those
moments have passed, though Their memories will always be with us.

- Id 7: In your smiles there was a peace that we have never found in
anything else, in your silences there was a perfection that we have
never seen in our lives. Until a man understands another, he realizes
his own reality. Can't understand. Your memories never gave us a chance
to rest, but your closeness changed our destiny.

- Id 8: Thousands of wishes such that every wish came out, my dreams
came out a lot, but still few came out, there was nothing, at least it
was that someone appeared, the secret of life in the beating of hearts
sometimes came out less.

- Id 9: Know yourself! This is your search for truth, this is your
realization of the secret of the world and the purpose of life, this is
the voice of your self, which prompts you to reach this truth. Man can
fulfill himself only by knowing his own reality, and when he knows his
own reality, he not only changes his own destiny but also has the
ability to change the world. This is the secret that no wisdom, no
philosophy, no book can explain. This can only be understood by one who
realizes this reality within himself.

- Id 10: There was nothing of the state of the heart, to understand the
state of the heart, there was nothing of the reality of sight, to
understand the reality of the sight. It is man's act of recognizing the
truth within himself that gives him the power to change his destiny. Man
has not come to this world only for self-gratification, but his purpose
is to recognize his natural abilities and make the best use of them. The
purpose of life is not only to find happiness and peace, but to go
beyond that to improve one's character and help others. Man cannot
change his destiny unless he understands the truth within himself and
acts on it. Its purpose is to raise one's ego and bring maturity in
one's intentions. When a person balances his thoughts and actions, he
not only improves his life but also becomes an example for others. Such
a person finds himself true and successful in the world, because he
always lives according to reality.

- Id 11: Thousands of wishes such that every wish came out, my dreams
came out a lot but still few came out. There was nothing, the least was
that someone appeared, the secret of life in the beating of hearts never
came out less.

- Id 12: People of our nation spend time that they should not take care
of their health, but in every gathering there is no time to comment on
the health of others. It seems like everyone is born a doctor.

- Id 13: Love is a selfless emotion that teaches a person to go beyond
the limits of his own self and rejoice in the happiness of others. Love
is an act in which giving is more enjoyable than receiving. Those who
understand love, they find the real secret of life, because love is the
greatest blessing of God.

- Id 14: The purpose of education is not only to get employment but to
shape the character of a person, to elevate his thinking and to make him
a means of improvement in the society.

- Id 15: Knowledge is the greatest wealth of man, which not only shows
man the ways of success in his life but also enables him to help others.
By lighting the lamp of knowledge, man moves from darkness to light.

- Id 16: Love is a disease and until it lasts, it is not a disease. Just
a joke.

- Id 17: There is no person in the world who loves someone more than his
own existence, his own self, we are all selfish. Why would anyone make a
relationship with?

- Id 18: Studying without a purpose is not only useless but also
impossible. The more we read aimlessly, the more we drift away from
meaningless matters.

- Id 19: The truth is that only fools know the joy of laziness. People
who run around, get up early in the morning and like exercise, what do
they know about this fun?

- Id 20: When will the day come when Hindu students will call Anis the
greatest poet and Muslim students will consider Tulsidas ten times
greater than Anis? A Hindu student will die for an Urdu ghazal and a
Muslim student will swear by Mahadevi's soft, juicy and heart touching
songs.

###Checks

 Adding quotes?

- [x] Make sure to include translations for the quotes in the
description (or another comment) so we can verify their content.
- [x] Check if any open issues are related to this PR; if so, be sure to
tag them below.
- [x] Make sure the PR title follows the Conventional Commits standard.
(https://www.conventionalcommits.org/ for more info)
- [x] Make sure to include your GitHub username prefixed with @ inside
parentheses at the end of the PR title.