fix(language): swiss german 1k typos (@boessu) (#6171)
fix: Correcting spelling and formal errors 

### Description
Correcting spelling errors (according to Hunspell swiss german corrector
in LibreOffice) and replacing phrases with corresponding words.

By the way: the link in the "_comment" is a dead link. I'm not sure if
that was somewhat a living link in the past. You won't find it in the
Wayback Machine (archive.org) either.

### Checks
- [ ] Adding quotes?
- [ ] Make sure to include translations for the quotes in the
description (or another comment) so we can verify their content.
- [ ] Adding a language or a theme?
- [ ] If is a language, did you edit `_list.json`, `_groups.json` and
add `languages.json`?
  - [ ] If is a theme, did you add the theme.css?
- Also please add a screenshot of the theme, it would be extra awesome
if you do so!
- [x] Check if any open issues are related to this PR; if so, be sure to
tag them below.
- [x] Make sure the PR title follows the Conventional Commits standard.
(https://www.conventionalcommits.org for more info)
- [x] Make sure to include your GitHub username prefixed with @ inside
parentheses at the end of the PR title.