impr(quote): add german quote (@boessu) (#6175)
feat: Add a quote of Patent Ochsner in the file.

### Description

Just add one quote from Patent Ochsner to lighten up everybody's day
(translation):
"A glass to love and one to the full life. And one to that which we
cannot hold. One gate opens and another closes. You stay in my heart,
even when it starts to hurt."

### Checks

- [x] Adding quotes?
- [x] Make sure to include translations for the quotes in the
description (or another comment) so we can verify their content.
- [ ] Adding a language or a theme?
- [ ] If is a language, did you edit `_list.json`, `_groups.json` and
add `languages.json`?
  - [ ] If is a theme, did you add the theme.css?
- Also please add a screenshot of the theme, it would be extra awesome
if you do so!
- [x] Check if any open issues are related to this PR; if so, be sure to
tag them below.
- [x] Make sure the PR title follows the Conventional Commits standard.
(https://www.conventionalcommits.org for more info)
- [x] Make sure to include your GitHub username prefixed with @ inside
parentheses at the end of the PR title.