fix: Remove seldomly used punctuation mark from Greek lang
The `·` (U+00B7, middle dot or "ano teleia") is a valid punctuation mark
in Greek. However, it has fallen into disuse in contemporary times. Most
people know it's meaning but very rarely use it in texts. They opt for
period (`.`) or commna(`,`) depending on context.

There's a practical reason for removing it too. You can't easily input it.
There isn't a dedicated key on a keyboard for it.
On Windows, you have to use `Alt`+`0183` AND have a numpad. If you have eg a 75%
keyboard you can't use that.