impr(quotes): add Urdu quotes (@aitchz) (#6427)
### Description
In this pull request, I have added a set of Urdu quotes to contribute to
the collection of quotes in Monkeytype.

### Translation

- **ID 21**: When man attains self-knowledge, he is able to understand
the realities of the world. Real humanity is to recognize the light
within and use that light for the benefit of others. This world is a
mirror that shows us the reflection of our mistakes and weaknesses so
that we can improve ourselves. Those who adjust their thinking and
actions according to higher principles, those are the ones who are truly
successful in life. Knowledge and character are two treasures that never
expire, and these two things take a man to the heights of greatness.

- **ID 22**: The greatest obstacle in the way of success is man himself.
If man recognizes his weaknesses and overcomes them, nothing can stop
him from becoming successful. Always remember that the acquisition of
knowledge and its Action is the real success. Those who limit knowledge
only to themselves, they cannot bring any positive change in the world
are successful.


### Checks

- [x] Adding quotes?
- [x] Make sure to include translations for the quotes in the
description (or another comment) so we can verify their content.
- [x] Check if any open issues are related to this PR; if so, be sure to
tag them below.
- [x] Make sure the PR title follows the Conventional Commits standard.
(https://www.conventionalcommits.org for more info)
- [x] Make sure to include your GitHub username prefixed with @ inside
parentheses at the end of the PR title.