feat(layout): add OPY (@robi-wan) (#6453)
### Description

Add OPY layout

https://github.com/voidyourwarranty2/opy-layout
> OPY is a keyboard layout in the tradition of NEO2, AdNW and KOY
> for a split ortholinear keyboard with shift and space on thumb keys,
> optimized for mixed English and German usage [..]

I tweaked the rows so that it looks good in the matrix and split matrix
keyboard styles. The author did not define a row1 so I took the freedom
to include one which resembles row1 from the qwertz layout.

### Checks

- [ ] Adding quotes?
- [ ] Make sure to include translations for the quotes in the
description (or another comment) so we can verify their content.
- [ ] Adding a language or a theme?
- [ ] If is a language, did you edit `_list.json`, `_groups.json` and
add `languages.json`?
  - [ ] If is a theme, did you add the theme.css?
- Also please add a screenshot of the theme, it would be extra awesome
if you do so!
- [x] Check if any open issues are related to this PR; if so, be sure to
tag them below.
- [x] Make sure the PR title follows the Conventional Commits standard.
(https://www.conventionalcommits.org for more info)
- [x] Make sure to include your GitHub username prefixed with @ inside
parentheses at the end of the PR title.

---------

Co-authored-by: Christian Fehmer <cfe@sexy-developer.com>
Co-authored-by: Miodec <jack@monkeytype.com>