impr(quotes): 15 polish proverbs and other quotes for polish language (@sobczaktm) (#6760)
Description
15 polish proverbs and other quotes for polish language

Checks
Translations:
Agreement is reached when one party pretends to believe the nonsense
told by the other. - "id":216

Only six people in the Galaxy knew that the president's job was not to
wield power, but to distract from it. - "id":217

I have become death, the destroyer of worlds. - "id":218

If something is impossible to do, let the Poles do it.- "id":219

It's not how long you live, but how wisely you live. It's never too late
to learn and grow. Wisdom is priceless, regardless of age. - "id":220

Don't waste time, for it is the stuff life is made of. - "id":221

You praise others, you do not know your own, you do not know what you
have. - "id":222

Don't argue with a fool because he will drag you down to his level and
beat you with experience. - "id":223

It is better to remain silent and be thought a fool than to open your
mouth and remove all doubt. - "id":224

Clothes do not make the man, clothes are for the man, not man for
clothes. -"id":225

He who digs holes for another will fall into them himself. Therefore, do
not do to others what you would not want done to you. - "id":226

Don't judge a man by his clothes, but by the friends he has. For a man
is what he associates with. - "id":227

What goes around comes around. Don't put off until tomorrow what you can
do today. - "id":228

First a person learns to walk, then to run, then to speak, then to be
silent. It's crucial to do certain things in the right order. - "id":229

Don't praise the day before sunset, you're not sure what the end of the
day may bring.- "id":230

---------

Co-authored-by: Tomasz.Sobczak <Tomasz.Sobczak@fujitsu.com>