fix(account-history): highlight animation is not working on result selection (@byseif21) (#6744)
### Description

fix the highlight animation for the account history table (when clicking
a chart dot) was not visible or only worked inconsistently. This was
because the animation was applied to the <tr> element, but the
background color for table rows is set on the <td> elements. As a
result, the animation was hidden by the static background color of the
table cells.

Closes #

---------

Co-authored-by: Miodec <jack@monkeytype.com>