fix(funbox): fix animations for choo_choo and earthquake funboxes in custom mode (@ShizukoV) (#6815)
Fixed an issue where the `choo_choo` and `earthquake` funboxes did not
"work" in custom mode.

Previously, these funboxes did not "work" in custom mode. When they were
turned on, *no animations were played*. This change allows the
animations to run regardless of the mode, so users can now experience
the funboxes in custom mode as well. This change also makes sure they
work on the other modes.

### Here is a video example of the bugs and the new changes in action:


[![funboxesFixedVideo](https://img.youtube.com/vi/oF1zuqWGYnQ/0.jpg)](https://www.youtube.com/watch?v=oF1zuqWGYnQ)