fix(lazy-mode): Save lazy mode in local storage after getting not supported message (@Leonabcd123) (#7171)
### Description

Save lazy mode status in local storage after getting the "This language
does not support lazy mode" message. This is so the user doesn't get the
same message every time they refresh, and it matches up with the
settings that show that lazy mode is disabled even though it's enabled
in local storage.